Extract out faiss index
